requirements.txt
asgiref==3.7.2
Django==4.2.5
django-otp==1.2.3
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
